+++ tests/dummy/app/controllers/radio.js
@@ -1,0 +1,10 @@
+import Ember from 'ember';
+
+export default Ember.Controller.extend({
+
+  actions: {
+    foo() {
+
+    }
+  }
+});